Another horrific code refactor
