+++ viewer/management/commands/generate_walk_events.py
@@ -2,7 +2,7 @@ from django.core.management.base import BaseCommand
 from django.core.cache import cache
 from django.conf import settings
 from viewer.models import DataReading, Event, EventWorkoutCategory
-from viewer.tasks import regenerate_similar_events
+from viewer.tasks.datacrunching import regenerate_similar_events
 from viewer.functions.geo import get_area_name
 from viewer.functions.location_manager import getgeoline, get_location_manager_report_queue
 from xml.dom import minidom
