+++ viewer/tasks/__init__.py
@@ -1,3 +1,4 @@
+from viewer.tasks.imports import *
 from viewer.tasks.process import *
 from viewer.tasks.reports import *
 from viewer.tasks.datacrunching import *
