+++ viewer/views/events.py
@@ -6,9 +6,10 @@ from django.db.models import Q, F
 from django.conf import settings
 import datetime, pytz, dateutil.parser, json, requests, random
 
-from viewer.tasks import regenerate_similar_events, generate_similar_events, generate_photo_collages
 from viewer.models import Person, Photo, Event, EventWorkoutCategory, CalendarAppointment
 from viewer.forms import EventForm, QuickEventForm
+from viewer.tasks.reports import generate_photo_collages
+from viewer.tasks.datacrunching import regenerate_similar_events, generate_similar_events
 
 from viewer.functions.moonshine import get_moonshine_tracks
 from viewer.functions.locations import join_location_events
@@ -38,10 +39,6 @@ def events(request):
 			raise Http404(form.errors)
 
 	data = {}
-	#data['event'] = Event.objects.filter(type='event', workout_categories=None).order_by('-start_time')[0:10]
-	#data['journey'] = Event.objects.filter(type='journey', workout_categories=None).order_by('-start_time')[0:10]
-	#data['workout'] = Event.objects.exclude(workout_categories=None).order_by('-start_time')[0:10]
-	#data['photo'] = Event.objects.filter(type='photo').exclude(caption='Photos').order_by('-start_time')[0:10]
 	data['life'] = Event.objects.filter(type='life_event').order_by('-start_time')
 	form = EventForm()
 	context = {'type':'view', 'data':data, 'form':form, 'categories':EventWorkoutCategory.objects.all()}
